+++ tests/index.test.js
@@ -3,7 +3,7 @@
 
 const assert = require('assert');
 
-const { parseXml, XmlCdata, XmlComment, XmlDeclaration, XmlDocument, XmlDocumentType, XmlElement, XmlNode, XmlProcessingInstruction, XmlText } = require('..');
+const { parseXml, XmlCdata, XmlComment, XmlDeclaration, XmlDocument, XmlDocumentType, XmlElement, XmlError, XmlNode, XmlProcessingInstruction, XmlText } = require('..');
 
 it('exports XML node classes', () => {
   assert.equal(typeof XmlCdata, 'function');
@@ -12,6 +12,7 @@ it('exports XML node classes', () => {
   assert.equal(typeof XmlDocument, 'function');
   assert.equal(typeof XmlDocumentType, 'function');
   assert.equal(typeof XmlElement, 'function');
+  assert.equal(typeof XmlError, 'function');
   assert.equal(typeof XmlNode, 'function');
   assert.equal(typeof XmlProcessingInstruction, 'function');
   assert.equal(typeof XmlText, 'function');